Rename diary_dir config key to dir
This renames the config key diary_dir to dir and changes a few
references in the code along the line, for instance, the key 'dir' in
the global CONFIG struct.

This way, the cmd line option is the same as the key in the config file.

For the environment variable, it is best to leave it at $DIARY_DIR,
because $DIR alone is certainly too unspecific.