Reorder options and precedence example
This commit reorders the options and arguments in the help texts:
* version
* help
* dir
* editor
* fmt
* range
* weekday

Further, it provides an example on variable precedence. The precedence
is roughly defined as follows:
1. Default values in the source code
2. Config file
3. Environment variables
4. Option arguments
5. Non-option argument

To make this order consistent for the weekday value derived from the
glibc locale, the commit includes an additional check for the $LANG
environment variable.