fix segfault after oauth
The global var access_token did not get initialized the first time it
was used in a call to caldav_req(), because the tokenfile was only
re-read on a call to get_access_token() with refresh=true.

This flag refresh=true is used to refresh with the refresh token from
the tokenfile. The global vars access_token and refresh_token however
should be re-read from the token file in both case, refres=true and
refresh=false.