discard Pad for oauth href
Rendering the href for oauth on a Pad does not solve the problem of the
non clickable link in the terminal emulator. Even though, the pad can
hold the full link in a single line and we increase the Pad width to
inf or 9999, the clickable part of the link will only go as far as the
max width of the window. Tested with Konsole.